// src/common/enum/notification_type.enum.ts
export enum NotificationType {
  ORDER_CREATED = 'ORDER_CREATED',
  ORDER_UPDATED = 'ORDER_UPDATED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  PAYMENT_RECEIVED = 'PAYMENT_RECEIVED',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  LOW_STOCK = 'LOW_STOCK',
  NEW_USER = 'NEW_USER',
  SYSTEM_ALERT = 'SYSTEM_ALERT',
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  LOGOUT = 'LOGOUT',
  PASSWORD_RESET = 'PASSWORD_RESET',
  ACCOUNT_BLOCKED = 'ACCOUNT_BLOCKED',
  ACCOUNT_DELETED = 'ACCOUNT_DELETED',
  GOOGLE_OAUTH = 'GOOGLE_OAUTH',
  CUSTOM = 'CUSTOM',
}